Laravue_19_09_2024

+++ app/Http/Controllers/ListeingImageController.php
@@ -5,13 +5,13 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ImageRequest;
 use App\Models\Listeing;
 use App\Models\ListeingImage;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
 class ListeingImageController extends Controller
 {
     public function create(Listeing $listeing)
     {
+        $this->authorize('delete', $listeing);
         return inertia('Realtor/ListeingImage/Create', [
             'listeing' => $listeing->load('images'),
         ]);
@@ -19,6 +19,7 @@ class ListeingImageController extends Controller
 
     public function store(ImageRequest $request, Listeing $listeing)
     {
+        $this->authorize('delete', $listeing);
         foreach ($request->images as $image) {
             $path = $image->store('images', 'public');
             $listeing->images()->create(['filename' => $path]);
@@ -29,6 +30,7 @@ class ListeingImageController extends Controller
 
     public function destroy(ListeingImage $image)
     {
+        $this->authorize('delete', $image->listeing);
         Storage::disk('public')->delete($image->filename);
         $image->delete();
